text_generator/requirements.txt
torch
diffusers
gradio
pillow
